refactor: apiからのデータ取得処理をapiディレクトリに配置 #4
