+++ src/app/routes/population-graph.tsx
@@ -10,7 +10,10 @@ import {
 } from "@/features/population-graph/types/prefectures";
 import styles from "@/features/population-graph/styles/contents.module.css";
 
-const url = "https://yumemi-frontend-engineer-codecheck-api.vercel.app/api/v1/";
+import {
+  getPrefectures,
+  getPopulationData,
+} from "@/features/population-graph/api/get-prefecture-data";
 
 const PopulationGraph = () => {
   const [prefectures, setPrefectures] = useState<RegionData[] | null>(null);
@@ -20,76 +23,17 @@ const PopulationGraph = () => {
   const [apiKey, setApiKey] = useState<string>("");
 
   const handleGetPrefectures = async () => {
-    if (!apiKey) return;
-
-    await fetch(url + "prefectures", {
-      method: "GET",
-      headers: {
-        "Content-Type": "application/json",
-        "x-api-key": apiKey,
-      },
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        console.log("成功:", data);
-        const initialPrefectures: RegionData[] = data.result.map(
-          (pref: RegionData) => ({
-            ...pref,
-            isSelected: false,
-          })
-        );
-        setPrefectures(initialPrefectures);
-        getPopulationData(initialPrefectures);
-      })
-      .catch((error) => {
-        console.error("エラー:", error);
-      });
-  };
-
-  const getPopulationData = async (prefectures: RegionData[]) => {
-    const tempPopulationData: PopulationData | null = {
-      boundaryYear: 0,
-      data: [],
-    };
-    prefectures?.forEach(async (pref) => {
-      await fetch(
-        url + "population/composition/perYear?prefCode=" + pref.prefCode,
-        {
-          method: "GET",
-          headers: {
-            "Content-Type": "application/json",
-            "x-api-key": apiKey,
-          },
-        }
-      )
-        .then((res) => res.json())
-        .then((data) => {
-          console.log("成功:", data);
-          tempPopulationData.boundaryYear = data.result.boundaryYear;
-          tempPopulationData.data =
-            tempPopulationData === null
-              ? [
-                  {
-                    prefCode: pref.prefCode,
-                    prefName: pref.prefName,
-                    ...data.result.data,
-                  },
-                ]
-              : [
-                  ...tempPopulationData.data,
-                  {
-                    prefCode: pref.prefCode,
-                    prefName: pref.prefName,
-                    ...data.result.data,
-                  },
-                ];
+    const data = await getPrefectures(apiKey);
+    if (data && data.result) {
+      const initialPrefectures: RegionData[] = data.result.map(
+        (pref: RegionData) => ({
+          ...pref,
+          isSelected: false,
         })
-        .catch((error) => {
-          console.error("エラー:", error);
-        });
-    });
-
-    setPopulationData(tempPopulationData);
+      );
+      setPrefectures(initialPrefectures);
+      getPopulationData(apiKey, initialPrefectures, setPopulationData);
+    }
   };
 
   return (
